docs: Add file to generate sample graph
This is the "dot" file that is used to generate assets/sample_graph.svg.
I'm not committing the other input files because it would be 10ish files
just for some sample data, but I do want to include this file so I can
generate the SVG file later if necessary.